backend/requirements/base.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt===5.3.1
dj-rest-auth===5.0.2
django-allauth===0.61.1
gunicorn==21.2.0
kombu==5.3.4
packaging==23.2
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
pillow==10.2.0
django-resized==1.0.2
django-ckeditor==6.7.0